requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Data handling
pandas==2.2.2
numpy==1.26.4

# Reddit API
praw==7.7.1
python-dotenv==1.0.1

# NLP / Hugging Face
transformers==4.44.2
torch==2.4.1
tokenizers==0.19.1

# Optional plotting / debugging
matplotlib==3.9.2
plotly==5.23.0

# Development extras (remove in production if not needed)
watchfiles==0.22.0  # for --reload in uvicorn
